clean up - on to FOL
