+++ common/engine/src/main/kotlin/com/benjishults/bitnots/engine/proof/strategy/closing_strategies.kt
@@ -20,11 +20,11 @@ interface BranchCloser
 
 data class PropositionalBranchCloser(val pos: SignedFormula<*>? = null, val neg: SignedFormula<*>? = null) : BranchCloser
 
-interface NodeClosingStrategy {
+interface TableauClosingStrategy {
     fun populateBranchClosers(tableau: Tableau)
 }
 
-object PropositionalNodeClosingStrategy : NodeClosingStrategy {
+object PropositionalClosingStrategy : TableauClosingStrategy {
     override fun populateBranchClosers(tableau: Tableau) {
         tableau.root.preorderIterator<TableauNode>().let { iter ->
             while (iter.hasNext()) {
@@ -76,7 +76,6 @@ object PropositionalNodeClosingStrategy : NodeClosingStrategy {
 
 interface InProgressTableauClosedIndicator {
     fun createExtension(closer: BranchCloser): InProgressTableauClosedIndicator
-    //    val needToClose: Stack<out TableauNode>
     val branchClosers: List<BranchCloser>
 
     /**
@@ -86,7 +85,7 @@ interface InProgressTableauClosedIndicator {
 
     fun progress(): InProgressTableauClosedIndicator
 
-    fun isCloser(): Boolean // = needToClose.isEmpty()
+    fun isCloser(): Boolean
 }
 
 /**
@@ -128,19 +127,23 @@ open class BooleanClosedIndicator private constructor(
         }
     }
 
-    override val branchClosers: List<BranchCloser> = branchClosers
+    override val branchClosers: List<BranchCloser> =
+            branchClosers
 
-    private lateinit var needToClose: Stack<TableauNode>// = nodes.clone() as Stack<TableauNode>
+    private lateinit var needToClose: Stack<TableauNode>
 
-    override fun isCloser() = needToClose.isEmpty()
+    override fun isCloser() =
+            needToClose.isEmpty()
 
-    override fun nextNode(): TableauNode = needToClose.peek()
+    override fun nextNode(): TableauNode =
+            needToClose.peek()
 
     override fun createExtension(closer: BranchCloser): InProgressTableauClosedIndicator {
         return BooleanClosedIndicator(branchClosers + closer, needToClose)
     }
 
     override fun progress(): InProgressTableauClosedIndicator =
+            @Suppress("UNCHECKED_CAST")
             (nextNode().children as List<TableauNode>).takeIf {
                 it.isNotEmpty()
             }?.reversed()?.let {
@@ -149,16 +152,12 @@ open class BooleanClosedIndicator private constructor(
                     it.forEach {
                         push(it)
                     }
-                    push(orig) // pushing this on so that it will be popped by the constructor
+                    push(orig) // pushing this back on so that it will be popped by the constructor
                 }.let { newNeeds ->
                     BooleanClosedIndicator(branchClosers, newNeeds)
                 }
             } ?: NotCompatible
 
-}
-
-interface ClosingStrategy {
-    fun checkClosed(node: TableauNode): InProgressTableauClosedIndicator
 }
 
 //open class FolUnificationClosingStrategy : ClosingStrategy {
@@ -178,178 +177,6 @@ interface ClosingStrategy {
 //                }
 //            }
 //            pos.any { p -> neg.any { it === p } }.takeIf { it }?.let { Closed } ?: NotClosed
-//        }
-//    }
-//
-//}
-
-interface InitializingStrategy {
-    fun init(node: TableauNode)
-}
-
-interface PropositionalInitializingStrategy : InitializingStrategy {
-    override fun init(node: TableauNode) {
-        with(node) {
-            applyAllAlphas(node)
-            applyRegularity(node)
-            allFormulas.addAll(newFormulas)
-        }
-    }
-
-    fun applyAllAlphas(node: TableauNode)
-    fun applyRegularity(node: TableauNode)
-}
-
-open class PropositionalInitializationStrategy : PropositionalInitializingStrategy {
-
-    override fun applyAllAlphas(node: TableauNode) {
-        with(node) {
-            while (true) {
-                val toAdd: MutableList<SignedFormula<Formula<*>>> = mutableListOf()
-                newFormulas.iterator().let {
-                    while (it.hasNext()) {
-                        val current = it.next()
-                        if (current is AlphaFormula) {
-                            it.remove()
-                            toAdd.addAll(current.generateChildren());
-                        }
-                    }
-                }
-                if (toAdd.isEmpty())
-                    break
-                else
-                    newFormulas.addAll(toAdd)
-            }
-        }
-    }
-
-    override fun applyRegularity(node: TableauNode) {
-        with(node) {
-            newFormulas.iterator().let { iter ->
-                while (iter.hasNext()) {
-                    iter.next().let {
-                        if (it in allFormulas)
-                            iter.remove()
-                    }
-                }
-            }
-        }
-    }
-
-}
-
-interface StepStrategy<in T : Tableau> {
-    fun step(tableau: T): Boolean
-}
-
-open class PropositionalStepStrategy<N : TableauNode>(
-        val nodeFactory: (MutableList<SignedFormula<*>>, N) -> N
-) : StepStrategy<Tableau> {
-    override open fun step(tableau: Tableau): Boolean =
-            with(tableau) {
-                applyBeta(tableau)// { fs, le -> PropositionalTableauNode(fs, le, closingStrategy, initStrategy) }
-            }
-
-    fun applyBeta(tableau: Tableau): Boolean =
-            with(tableau) {
-                var beta: BetaFormula<Formula<*>>? = null
-                val node = root.breadthFirst<TableauNode> {
-                    beta = it.newFormulas.firstOrNull { it is BetaFormula<*> } as BetaFormula<*>?
-                    beta !== null
-                }
-                if (node === null)
-                    false
-                else
-                    beta?.let { beta ->
-                        node.newFormulas.remove(beta);
-                        val leaves = node.allLeaves<N>()
-                        leaves.map { leaf ->
-                            beta.generateChildren().map {
-                                leaf.children.add(nodeFactory(mutableListOf(it), leaf))
-                            }
-                        }
-                        true
-                    } ?: false
-            }
-
-}
-
-//open class FolStepStrategy(
-//        nodeFactory: (MutableList<SignedFormula<*>>, FolTableauNode) -> FolTableauNode
-//) : PropositionalStepStrategy<FolTableauNode>(nodeFactory) {
-//    override open fun step(tableau: Tableau): Boolean =
-//            with(tableau) {
-//                if (!applyDelta(tableau) && !applyBeta(tableau)) {
-//                    if (unify(tableau).needToClose.isEmpty())
-//                        true
-//                    else
-//                        applyGamma(tableau)
-//                } else {
-//                    true
-//                }
-//            }
-//
-//    fun unify(tableau: Tableau): MultiBranchCloser {
-//        with(tableau) {
-//            root.breadthFirst<FolTableauNode> { node ->
-//                node.toString()
-//                true
-//            }
-//            return MultiBranchCloser()
-//        }
-//    }
-//
-//    // TODO make this splice
-//    private fun applyDelta(tableau: Tableau): Boolean {
-//        with(tableau) {
-//            var delta: DeltaFormula<out VarBindingFormula>? = null
-//            val node = root.breadthFirst<FolTableauNode> {
-//                delta = it.newFormulas.firstOrNull { it is DeltaFormula<*> } as DeltaFormula<*>?
-//                delta !== null
-//            }
-//            if (node === null)
-//                return false
-//            delta?.let {
-//                delta ->
-//                node.newFormulas.remove(delta);
-//                val leaves = node.allLeaves<FolTableauNode>()
-//                leaves.forEach<FolTableauNode> { leaf ->
-//                    delta.generateChildren().forEach {
-//                        leaf.children.add(nodeFactory(mutableListOf(it), leaf))
-//                        Unit
-//                    }
-//                }
-//                return true
-//            } ?: return false
-//        }
-//    }
-//
-//    // TODO make this splice
-//    private fun applyGamma(tableau: Tableau): Boolean {
-//        with(tableau) {
-//            var gammas: MutableSet<Pair<GammaFormula<*>, FolTableauNode>> =
-//                    TreeSet<Pair<GammaFormula<*>, FolTableauNode>>(
-//                            Comparator<Pair<GammaFormula<*>, FolTableauNode>> {
-//                                o1: Pair<GammaFormula<*>, FolTableauNode>?, o2: Pair<GammaFormula<*>, FolTableauNode>? ->
-//                                o1!!.first.numberOfApplications.compareTo(o2!!.first.numberOfApplications)
-//                            })
-//            root.breadthFirst<FolTableauNode> { node ->
-//                gammas.addAll(node.newFormulas.filterIsInstance<GammaFormula<*>>().map {
-//                    it to node
-//                }) // stop if we get to one that is ready to go
-//                        && gammas.first().first.numberOfApplications == 0
-//            }
-//            gammas.firstOrNull()?.let {
-//                (gamma, node) ->
-//                gamma.numberOfApplications++
-//                val leaves = node.allLeaves<FolTableauNode>()
-//                leaves.map { leaf ->
-//                    gamma.generateChildren().map {
-//                        leaf.children.add(nodeFactory(mutableListOf(it), leaf))
-//                    }
-//                }
-//                return true
-//            } ?: return false
 //        }
 //    }
 //
